auth_service/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
PyJWT==2.3.0
requests==2.26.0
djangorestframework-simplejwt==4.8.0
whitenoise==5.3.0